+++ WargameEngine/WargameEngine/WargameEngine/controller/INetSocket.h
@@ -10,7 +10,7 @@ public:
 	virtual void InitClient(const char* ip, unsigned short port = 0) = 0;
 	virtual void InitHost(unsigned short port = 0) = 0;
 
-	virtual bool SendData(const char * data, int len) const = 0;
+	virtual bool SendData(const char * data, size_t len) const = 0;
 	//Return -1 then error occurs, 0 then connection is closed by other side or number of bytes received.
-	virtual int RecieveData(char* data, int maxLength) = 0;
+	virtual int RecieveData(char* data, size_t maxLength) = 0;
 };
